+++ src/plugin/Services.php
@@ -1,23 +1,28 @@
 <?php
 namespace fruitstudios\colorit\plugin;
 
-use fruitstudios\colorit\services\Colours;
+use fruitstudios\colorit\services\Colors;
 use fruitstudios\colorit\services\Fields;
-use fruitstudios\colorit\services\FieldTemplates;
+use fruitstudios\colorit\services\Presets;
 
 trait Services
 {
     // Public Methods
     // =========================================================================
 
-    public function getFieldTemplates(): FieldTemplates
+    public function getPresets(): Presets
     {
-        return $this->get('fieldTemplates');
+        return $this->get('presets');
     }
 
-    public function getColours(): Colours
+    public function getColors(): Colors
     {
-        return $this->get('colours');
+        return $this->get('colors');
+    }
+
+    public function getColours(): Colors
+    {
+        return $this->getColors();
     }
 
     public function getFields(): Fields
@@ -31,9 +36,9 @@ trait Services
     private function _setPluginComponents()
     {
         $this->setComponents([
-            'colours' => Colours::class,
-            'colors' => Colours::class,
-            'fieldTemplates' => FieldTemplates::class,
+            'colors' => Colors::class,
+            'colours' => Colors::class,
+            'presets' => Presets::class,
             'fields' => Fields::class,
         ]);
     }