+++ src/web/twig/Extension.php
@@ -1,14 +1,21 @@
 <?php
 namespace fruitstudios\colorit\web\twig;
 
-use fruitstudios\colorit\Plugin;
+use fruitstudios\colorit\Colorit;
 
 use Craft;
 
 class Extension extends \Twig_Extension
 {
+    protected $colors;
+
     // Public Methods
     // =========================================================================
+
+    public function __construct()
+    {
+        $this->colors = Colorit::$plugin->getColors();
+    }
 
     public function getName(): string
     {
@@ -18,13 +25,11 @@ class Extension extends \Twig_Extension
     public function getFilters(): array
     {
         return [
-            new \Twig_SimpleFilter('customFilter', [$this, 'customFilterFunction'])
+            new \Twig_SimpleFilter('hexIsWhite', [$this->colors, 'hexIsWhite']),
+            new \Twig_SimpleFilter('hexIsBlack', [$this->colors, 'hexIsBlack']),
+            new \Twig_SimpleFilter('hexIsTransparent', [$this->colors, 'hexIsTransparent']),
+            new \Twig_SimpleFilter('hexToRgb', [$this->colors, 'hexToRgb']),
+            new \Twig_SimpleFilter('hexToRgba', [$this->colors, 'hexToRgba']),
         ];
     }
-
-    public function customFilterFunction($string): string
-    {
-        return $string;
-    }
-
 }